refactor: 공고 지원 커스텀 훅
