+++ src/apis/fileApi.js
@@ -2,7 +2,6 @@ const postFile = async (data, token) => {
   const res = await fetch(import.meta.env.VITE_APP_API_URL + '/api/files', {
     method: 'POST',
     headers: {
-      'content-type': 'multipart/form-data',
       'authorization': 'Bearer ' + token
     },
     body: data
